fix: Added new structure to Facebook icon
