+++ src/components/index.ts
@@ -147,7 +147,9 @@ export * from './social-icons/Github';
 export * from './social-icons/Medium';
 export * from './social-icons/Reddit';
 export * from './social-icons/Twitter';
-export * from './social-icons/Facebook';
+export * from './social-icons/Facebook/Facebook';
+export * from './social-icons/Facebook/Facebook18';
+export * from './social-icons/Facebook/Facebook24';
 export * from './social-icons/Linkedin/Linkedin';
 export * from './social-icons/Linkedin/Linkedin18';
 export * from './social-icons/Linkedin/Linkedin24';
