--- src/components/social-icons/Facebook.tsx
@@ -1,46 +1,0 @@
-import React, { FC } from 'react';
-import { ISocialIconProps } from './type';
-
-export const IconFacebook: FC<ISocialIconProps> = ({
-	size = 16,
-	color = 'currentColor',
-}) => {
-	const size18 = (
-		<svg
-			width='9'
-			height='16'
-			viewBox='0 0 9 16'
-			fill='none'
-			xmlns='http://www.w3.org/2000/svg'
-		>
-			<path
-				d='M8.5 1.5C8.5 0.947715 8.05228 0.5 7.5 0.5H6.25C5.25544 0.5 4.30161 0.895088 3.59835 1.59835C2.89509 2.30161 2.5 3.25544 2.5 4.25V6.5H1.25C0.697715 6.5 0.25 6.94772 0.25 7.5V8.5C0.25 9.05229 0.697715 9.5 1.25 9.5H2.5V14.5C2.5 15.0523 2.94772 15.5 3.5 15.5H4.5C5.05228 15.5 5.5 15.0523 5.5 14.5V9.5H6.96922C7.42809 9.5 7.82807 9.1877 7.93937 8.74254L8.18937 7.74254C8.34715 7.11139 7.86979 6.5 7.21922 6.5H5.5V4.25C5.5 4.05109 5.57902 3.86032 5.71967 3.71967C5.86032 3.57902 6.05109 3.5 6.25 3.5H7.5C8.05228 3.5 8.5 3.05228 8.5 2.5V1.5Z'
-				fill={color}
-			/>
-		</svg>
-	);
-
-	const size24 = (
-		<svg
-			width='11'
-			height='20'
-			viewBox='0 0 11 20'
-			fill='none'
-			xmlns='http://www.w3.org/2000/svg'
-		>
-			<path
-				d='M11 1C11 0.447715 10.5523 0 10 0H8C6.67392 0 5.40215 0.526784 4.46447 1.46447C3.52678 2.40215 3 3.67392 3 5V8H1C0.447715 8 0 8.44771 0 9V11C0 11.5523 0.447715 12 1 12H3V19C3 19.5523 3.44772 20 4 20H6C6.55228 20 7 19.5523 7 19V12H9.21922C9.67809 12 10.0781 11.6877 10.1894 11.2425L10.6894 9.24253C10.8472 8.61139 10.3698 8 9.71922 8H7V5C7 4.73478 7.10536 4.48043 7.29289 4.29289C7.48043 4.10536 7.73478 4 8 4H10C10.5523 4 11 3.55228 11 3V1Z'
-				fill={color}
-			/>
-		</svg>
-	);
-
-	switch (size.toString()) {
-		case '16':
-			return size18;
-		case '24':
-			return size24;
-		default:
-			return size18;
-	}
-};
