+++ src/components/social-icons/Facebook/Facebook18.tsx
@@ -1,0 +1,20 @@
+import React, { FC } from 'react';
+import { ISocialIconProps } from '../type';
+
+export const IconFacebook18: FC<ISocialIconProps> = ({
+	size = 18,
+	color = 'currentColor',
+}) => (
+	<svg
+		width={size}
+		height={size}
+		viewBox='0 0 18 18'
+		fill='none'
+		xmlns='http://www.w3.org/2000/svg'
+	>
+		<path
+			d='M13.5 2.5C13.5 1.94772 13.0523 1.5 12.5 1.5H11.25C10.2554 1.5 9.30161 1.89509 8.59835 2.59835C7.89509 3.30161 7.5 4.25544 7.5 5.25V7.5H6.25C5.69772 7.5 5.25 7.94772 5.25 8.5V9.5C5.25 10.0523 5.69772 10.5 6.25 10.5H7.5V15.5C7.5 16.0523 7.94772 16.5 8.5 16.5H9.5C10.0523 16.5 10.5 16.0523 10.5 15.5V10.5H11.9692C12.4281 10.5 12.8281 10.1877 12.9394 9.74254L13.1894 8.74254C13.3472 8.11139 12.8698 7.5 12.2192 7.5H10.5V5.25C10.5 5.05109 10.579 4.86032 10.7197 4.71967C10.8603 4.57902 11.0511 4.5 11.25 4.5H12.5C13.0523 4.5 13.5 4.05228 13.5 3.5V2.5Z'
+			fill={color}
+		/>
+	</svg>
+);
